requirements.txt
django-extensions==1.6.7
jsonfield==1.0.3
requests==2.9.1
django_allauth==0.25.2
python-twitter==3.1
facepy==1.0.8
hashids==1.1.0
